Fix Elasticsearch document index
