+++ src/OrchardCore/OrchardCore.Search.Elasticsearch.Core/Services/ElasticsearchIndexManager.cs
@@ -1,6 +1,7 @@
 using System.Collections.Concurrent;
 using System.Text.Json;
 using System.Text.Json.Nodes;
+using Azure;
 using Elastic.Clients.Elasticsearch;
 using Elastic.Clients.Elasticsearch.Analysis;
 using Elastic.Clients.Elasticsearch.Fluent;
@@ -224,7 +225,7 @@ public sealed class ElasticsearchIndexManager
                         {
                             [nameof(ContainedPartModel.Ids)] = new KeywordProperty(),
                             [nameof(ContainedPartModel.Order)] = new FloatNumberProperty(),
-                        }
+                        },
                     },
 
                     // We map DisplayText here because we have 3 different fields with it.
@@ -236,7 +237,7 @@ public sealed class ElasticsearchIndexManager
                             [nameof(DisplayTextModel.Analyzed)] = new TextProperty(),
                             [nameof(DisplayTextModel.Normalized)] = new KeywordProperty(),
                             [nameof(DisplayTextModel.Keyword)] = new KeywordProperty(),
-                        }
+                        },
                     },
 
                     // We map ContentType as a keyword because else the automatic mapping will break the queries.
@@ -485,20 +486,32 @@ public sealed class ElasticsearchIndexManager
         return indexName;
     }
 
-    public Task StoreDocumentsAsync(string indexName, IEnumerable<DocumentIndex> indexDocuments)
+    public async Task StoreDocumentsAsync(string indexName, IEnumerable<DocumentIndex> indexDocuments)
     {
         ArgumentException.ThrowIfNullOrEmpty(indexName);
         ArgumentNullException.ThrowIfNull(indexDocuments);
 
         if (indexDocuments == null || !indexDocuments.Any())
         {
-            return Task.CompletedTask;
+            return;
         }
 
-        return _elasticClient.BulkAsync(GetFullIndexNameInternal(indexName), descriptor => descriptor
-            .IndexMany(indexDocuments.Select(CreateElasticDocument))
-            .Refresh(Refresh.True)
-        );
+        var indexFullName = GetFullIndexNameInternal(indexName);
+
+        var response = await _elasticClient.BulkAsync(indexFullName,
+            descriptor => descriptor.CreateElasticDocument(indexDocuments).Refresh(Refresh.True));
+
+        if (!response.IsValidResponse)
+        {
+            if (response.TryGetOriginalException(out var ex))
+            {
+                _logger.LogWarning("There were issues indexing a document using Elasticsearch. Exception: {OriginalException}", ex);
+            }
+            else
+            {
+                _logger.LogWarning("There were issues indexing a document using Elasticsearch.");
+            }
+        }
     }
 
     public async Task<ElasticsearchResult> SearchAsync(ElasticsearchSearchContext context)
@@ -607,81 +620,6 @@ public sealed class ElasticsearchIndexManager
         }
     }
 
-    private static Dictionary<string, object> CreateElasticDocument(DocumentIndex documentIndex)
-    {
-        var entries = new Dictionary<string, object>
-        {
-            { IndexingConstants.ContentItemIdKey, documentIndex.ContentItemId },
-            { IndexingConstants.ContentItemVersionIdKey, documentIndex.ContentItemVersionId }
-        };
-
-        foreach (var entry in documentIndex.Entries)
-        {
-            switch (entry.Type)
-            {
-                case DocumentIndexBase.Types.Boolean:
-                    if (entry.Value is bool boolValue)
-                    {
-                        AddValue(entries, entry.Name, boolValue);
-                    }
-                    break;
-
-                case DocumentIndexBase.Types.DateTime:
-
-                    if (entry.Value is DateTimeOffset offsetValue)
-                    {
-                        AddValue(entries, entry.Name, offsetValue);
-                    }
-                    else if (entry.Value is DateTime dateTimeValue)
-                    {
-                        AddValue(entries, entry.Name, dateTimeValue.ToUniversalTime());
-                    }
-
-                    break;
-
-                case DocumentIndexBase.Types.Integer:
-                    if (entry.Value != null && long.TryParse(entry.Value.ToString(), out var value))
-                    {
-                        AddValue(entries, entry.Name, value);
-                    }
-
-                    break;
-
-                case DocumentIndexBase.Types.Number:
-                    if (entry.Value != null)
-                    {
-                        AddValue(entries, entry.Name, Convert.ToDouble(entry.Value));
-                    }
-                    break;
-
-                case DocumentIndexBase.Types.Text:
-                    if (entry.Value != null)
-                    {
-                        var stringValue = Convert.ToString(entry.Value);
-
-                        if (!string.IsNullOrEmpty(stringValue))
-                        {
-                            AddValue(entries, entry.Name, stringValue);
-                        }
-                    }
-                    break;
-                case DocumentIndexBase.Types.GeoPoint:
-                    if (entry.Value is DocumentIndexBase.GeoPoint point)
-                    {
-                        AddValue(entries, entry.Name, GeoLocation.LatitudeLongitude(new LatLonGeoLocation
-                        {
-                            Lat = (double)point.Latitude,
-                            Lon = (double)point.Longitude,
-                        }));
-                    }
-
-                    break;
-            }
-        }
-
-        return entries;
-    }
-
     public string GetFullIndexName(string indexName)
     {
         ArgumentException.ThrowIfNullOrEmpty(indexName);
@@ -722,33 +660,6 @@ public sealed class ElasticsearchIndexManager
 
     private string GetFullIndexNameInternal(string indexName)
         => GetIndexPrefix() + _separator + indexName;
-
-    private static void AddValue(Dictionary<string, object> entries, string key, object value)
-    {
-        if (entries.TryAdd(key, value))
-        {
-            return;
-        }
-
-        // At this point, we know that a value already exists.
-        if (entries[key] is List<object> list)
-        {
-            list.Add(value);
-
-            entries[key] = list;
-
-            return;
-        }
-
-        // Convert the existing value to a list of values.
-        var values = new List<object>()
-        {
-            entries[key],
-            value,
-        };
-
-        entries[key] = values;
-    }
 
     private string GetIndexPrefix()
     {
